NSX-T: Describe setting up Load Balancer
- we broke the "Setting up NSX-T" into several sections;
  our additions would have been overwhelming for a single
  section
- our naming is opinionated and differs from the NSX-T Cookbook's naming;
  - we don't refer to internal environments (e.g. `haas-108-gorouters`)
  - we don't conflate the VIP concepts with separate NSX-T concepts
     (`...gorouter-vip-80` vs. `pcf-web-vs-80`)
  - we refer to the virtual servers from the operator's perspective
    (`pcf-web-xxx`) rather than referring to PCF internal component
    name (`gorouter-`).
- also delete trailing whitespace

We have misgivings about the description of the DNS record of the
foundation. Rather than `*.YOUR-SYSTEM-DOMAIN.com`, we would have
preferred `*.your.apps.domain.tld`. Not everyone, not even Pivotal, ends
in `.com`. We also would have liked seeing a distinction between the
system domain and the apps domain (almost all the examples exclusively
used the system domain)

We would have welcomed a description of setting up the SNAT rule for
the VMs in the deployment; however, we were not sure if that lies
outside the scope of this documentation.

[#162096405](https://www.pivotaltracker.com/story/show/162096405)

Signed-off-by: Brian Cunnie <bcunnie@pivotal.io>